requirements.txt
google-cloud
pyinstaller
streamlit>=1.28.0
torch>=2.0.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Document loaders
pypdf>=3.17.0
langchain-google-vertexai